+++ internal/request/services/profile_service.go
@@ -1,4 +1,4 @@
-package application
+package services
 
 import (
 	"demo-signserver/internal/repository/domain"
